app/preprocess.py
from PIL import Image
import numpy as np

IMG_SIZE = 128

def preprocess_pil(image: Image.Image) -> np.ndarray:
    image = image.convert("RGB").resize((IMG_SIZE, IMG_SIZE))
    arr = np.asarray(image).astype("float32") / 255.0
    arr = np.expand_dims(arr, axis=0)
    return arr